Refactor app.py
MODIFIED:
>app.py
- Removed endpoints /question and /answer
- Removed session handling

ADDED:
>__init__.py to initialize the classes
>answer.py
   - Repsonsible of the /answer endpoint
>question.py
  - Responsible of the /question endpoint
  - Also responsible for generating new questions
>session.py
   - Responsible for session handling
