+++ src/answer.py
@@ -1,0 +1,40 @@
+from src import app
+
+from src.question import next_question
+
+import argparse
+import random
+import data.data as data
+from flask import Flask, escape, request, jsonify, session
+import string
+
+
+
+@app.route('/answer', methods=['POST'])
+def answer():
+    try:
+        content = request.get_json()
+        language = content['language']
+        attribute_id = content['attribute_id']
+        response = content['response']
+    except TypeError as e:
+        return jsonify({'success': False,
+                        'message': 'Please supply "language", "attribute_id", and "response" in query'})
+    except KeyError as e:
+        return jsonify({'success': False,
+                        'message': 'Please supply "language", "attribute_id", and "response" in query'})
+    else:
+        if 'user' in session:
+            # access users session data
+            users[session['user']]
+
+        posterior = data.calculate_posterior(attribute_id, response)
+        new_building_classes = []
+        for _, (class_id, score) in posterior.iterrows():
+            new_building_classes.append({'class_id': class_id,
+                                         'class_name': data.building_classes[class_id],
+                                         'score': score})
+
+        return jsonify({'success': True,
+                        'new_question': next_question(),
+                        'building_classes': new_building_classes})
